+++ example/lib/tex_view_quiz_example.dart
@@ -158,11 +158,11 @@ class _TeXViewQuizExampleState extends State<TeXViewQuizExample> {
               ),
               backgroundColor: Colors.white,
             ),
-            // loadingWidgetBuilder: (context) {
-            //   return const Center(
-            //     child: CircularProgressIndicator(),
-            //   );
-            // },
+            loadingWidgetBuilder: (context) {
+              return const Center(
+                child: CircularProgressIndicator(),
+              );
+            },
           ),
           if (isWrong)
             const Padding(