+++ lib/flutter_tex.dart
@@ -1,7 +1,6 @@
 library;
 
 export 'package:flutter_tex/src/models/font.dart';
-export 'package:flutter_tex/src/models/rendering_engine.dart';
 export 'package:flutter_tex/src/styles/border.dart';
 export 'package:flutter_tex/src/styles/font_style.dart';
 export 'package:flutter_tex/src/styles/margin.dart';