+++ lib/src/models/rendering_engine.dart
@@ -1,7 +1,0 @@
-class TeXViewRenderingEngine {
-  final String name;
-
-  const TeXViewRenderingEngine.katex() : name = "katex";
-
-  const TeXViewRenderingEngine.mathjax() : name = "mathjax";
-}