+++ lib/src/utils/tex_rendering_server.dart
@@ -2,36 +2,37 @@ import 'dart:async';
 
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_tex/src/models/rendering_engine.dart';
 import 'package:webview_flutter_plus/webview_flutter_plus.dart';
 
-
-
 /// A rendering server for TeXView. This is backed by a [LocalhostServer] and a [WebViewControllerPlus].
-/// Make sure to call [run] before using the [controller].
-class TeXRederingServer {
+/// Make sure to call [run] before using the [webViewControllerPlus].
+class TeXRenderingServer {
+  static final WebViewControllerPlus webViewControllerPlus =
+      WebViewControllerPlus();
+  static final LocalhostServer _server = LocalhostServer();
+
   static RenderingEngineCallback? onPageFinished,
       onTapCallback,
       onTeXViewRenderedCallback;
-  static WebViewControllerPlus controller = WebViewControllerPlus();
-  static TeXViewRenderingEngine renderingEngine =
-      const TeXViewRenderingEngine.mathjax();
 
-  static LocalhostServer server = LocalhostServer();
-
-  static Future<void> run({int port = 0}) async {
-    await server.start(port: port);
-  }
-
-  static Future<void> initController() async {
+  static Future<void> start(
+      {int port = 0, Map mathJaxConfig = const {}}) async {
     var controllerCompleter = Completer<void>();
 
-    controller
-      ..setJavaScriptMode(JavaScriptMode.unrestricted)
-      ..setBackgroundColor(Colors.transparent)
-      ..loadFlutterAssetWithServer(
-          "packages/flutter_tex/js/${renderingEngine.name}/index.html",
-          server.port!)
+    await _server.start(port: port);
+
+    webViewControllerPlus
+      ..addJavaScriptChannel(
+        'OnTapCallback',
+        onMessageReceived: (onTapCallbackMessage) =>
+            onTapCallback?.call(onTapCallbackMessage.message),
+      )
+      ..addJavaScriptChannel(
+        'TeXViewRenderedCallback',
+        onMessageReceived: (teXViewRenderedCallbackMessage) =>
+            onTeXViewRenderedCallback
+                ?.call(teXViewRenderedCallbackMessage.message),
+      )
       ..setNavigationDelegate(
         NavigationDelegate(
           onPageFinished: (String url) {
@@ -47,23 +48,16 @@ class TeXRederingServer {
           }
         },
       )
-      ..addJavaScriptChannel(
-        'OnTapCallback',
-        onMessageReceived: (onTapCallbackMessage) =>
-            onTapCallback?.call(onTapCallbackMessage.message),
-      )
-      ..addJavaScriptChannel(
-        'TeXViewRenderedCallback',
-        onMessageReceived: (teXViewRenderedCallbackMessage) =>
-            onTeXViewRenderedCallback
-                ?.call(teXViewRenderedCallbackMessage.message),
-      );
+      ..setJavaScriptMode(JavaScriptMode.unrestricted)
+      ..setBackgroundColor(Colors.transparent)
+      ..loadFlutterAssetWithServer(
+          "packages/flutter_tex/core/flutter_tex.html", _server.port!);
 
     return controllerCompleter.future;
   }
 
   static Future<void> stop() async {
-    await server.close();
+    await _server.close();
   }
 }