+++ lib/src/views/tex_view_web.dart
@@ -1,20 +1,24 @@
-// ignore_for_file: avoid_web_libraries_in_flutter
-
+import 'dart:js_interop';
+import 'dart:ui_web';
 import 'package:flutter/material.dart';
-import 'package:flutter/widgets.dart';
 import 'package:flutter_tex/flutter_tex.dart';
 import 'package:flutter_tex/src/utils/core_utils.dart';
-import 'dart:html' as html;
-import 'dart:js' as js;
-import 'dart:ui_web' as ui;
+import 'package:web/web.dart';
+
+@JS('TeXViewRenderedCallback')
+external set teXViewRenderedCallback(JSFunction callback);
+
+@JS('OnTapCallback')
+external set onTapCallback(JSFunction callback);
+
+@JS('initWebTeXView')
+external void initWebTeXView(String viewId, String rawData);
 
 class TeXViewState extends State<TeXView> {
   String? _lastData;
   double widgetHeight = initialHeight;
   final String _viewId = UniqueKey().toString();
-
-  final TeXViewRenderingEngine renderingEngine =
-      TeXRederingServer.renderingEngine;
+  bool _isReady = false;
 
   @override
   Widget build(BuildContext context) {
@@ -30,41 +34,45 @@ class TeXViewState extends State<TeXView> {
 
   @override
   void initState() {
-    _initWebview();
-    super.initState();
-  }
-
-  void _initWebview() {
-    ui.platformViewRegistry.registerViewFactory(
+    platformViewRegistry.registerViewFactory(
         _viewId,
-        (int id) => html.IFrameElement()
-          ..src =
-              "assets/packages/flutter_tex/js/${renderingEngine.name}/index.html"
+        (int id) => HTMLIFrameElement()
+          ..src = "assets/packages/flutter_tex/core/flutter_tex.html"
           ..id = _viewId
           ..style.height = '100%'
           ..style.width = '100%'
           ..style.border = '0');
 
-    js.context['TeXViewRenderedCallback'] = (message) {
-      double viewHeight = double.parse(message.toString());
-      if (viewHeight != widgetHeight) {
-        setState(() {
-          widgetHeight = viewHeight;
-        });
-      }
-    };
+    onTapCallback = onTap.toJS;
+    teXViewRenderedCallback = onTeXViewRendered.toJS;
 
-    js.context['OnTapCallback'] = (id) {
-      widget.child.onTapCallback(id);
-    };
+    _isReady = true;
+
+    _initTeXView();
+    super.initState();
+  }
+
+  void onTap(JSString id) {
+    widget.child.onTapCallback(id.toString());
+  }
+
+  void onTeXViewRendered(JSNumber message) {
+    double viewHeight = double.parse(message.toString());
+    if (viewHeight != widgetHeight) {
+      setState(() {
+        widgetHeight = viewHeight;
+      });
+    }
   }
 
   void _initTeXView() {
-    var rawData = getRawData(widget);
+    if (!_isReady) {
+      return;
+    }
 
+    var rawData = getRawData(widget);
     if (rawData != _lastData) {
-      js.context.callMethod(
-          'initWebTeXView', [_viewId, rawData, renderingEngine.name]);
+      initWebTeXView(_viewId, rawData);
       _lastData = rawData;
     }
   }