+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## [4.1.0] - Apr 01, 2025
+* MathJax upgraded to version 3.2, a huge performance improvement.
+* Katex has been removed.
+* Minor API changes, please refer to the example before upgrade the package.
+
+## [4.0.14] - Jan 04, 2025
+* Minmum flutter SDK requirement is 3.27.x
+* Please refer to the [TeXView example](https://github.com/Shahxad-Akram/flutter_tex/blob/master/example/lib) before using this version.
+
 ## [4.0.13] - Jan 03, 2025
 * Minmum flutter SDK requirement is 3.27.x
 * Please refer to the [TeXView example](https://github.com/Shahxad-Akram/flutter_tex/blob/master/example/lib) before using this version.